# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "bloomfilter router",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:65221",
    "basePath": "/api/v1",
    "paths": {
        "/bloomfilter/add": {
            "post": {
                "description": "query single value",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Slave"
                ],
                "parameters": [
                    {
                        "description": "add values",
                        "name": "values",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.AddRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "bool"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "bool"
                        }
                    }
                }
            }
        },
        "/bloomfilter/apply": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "apply bloomfilter memory",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Slave"
                ],
                "parameters": [
                    {
                        "description": "alloc size",
                        "name": "allocSize",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.ApplyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code\":2000,\"data\":null,\"message\":\"\"}",
                        "schema": {
                            "$ref": "#/definitions/handler.ApplyRes"
                        }
                    },
                    "400": {
                        "description": "code\":4000,\"data\":null,\"message\":\"\"}",
                        "schema": {
                            "$ref": "#/definitions/handler.ApplyRes"
                        }
                    }
                }
            }
        },
        "/bloomfilter/del/{uuid}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "delete key",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Slave"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "key(uuid)",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            }
        },
        "/bloomfilter/query": {
            "get": {
                "description": "query single value",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Slave"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "apply key",
                        "name": "key",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "query value",
                        "name": "value",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "bool"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "bool"
                        }
                    }
                }
            },
            "post": {
                "description": "query single value",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Slave"
                ],
                "parameters": [
                    {
                        "description": "add values",
                        "name": "values",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.AddRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "bool"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "bool"
                        }
                    }
                }
            }
        },
        "/bloomfilter/{address}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "backup address",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Slave"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "address(ip:port)",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.AddRequest": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "strings": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "app.ApplyRequest": {
            "type": "object",
            "properties": {
                "expiration": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "handler.ApplyRes": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "yes": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}